backend/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Werkzeug==0.9.6
distribute==0.7.3
wsgiref==0.1.2
Flask-SQLalchemy==2.0.0
